@@ -2,5 +2,4 @@ from store.models.base import CreateBaseModel
 from store.schemas.product import ProductIn
 
 
-class ProductModel(ProductIn, CreateBaseModel):
-    ...
+class ProductModel(ProductIn, CreateBaseModel): ...
