@@ -1,5 +1,4 @@
 from typing import List
-
 import pytest
 from tests.factories import product_data
 from fastapi import status
@@ -7,13 +6,11 @@ from fastapi import status
 
 async def test_controller_create_should_return_success(client, products_url):
     response = await client.post(products_url, json=product_data())
-
     content = response.json()
-
     del content["created_at"]
     del content["updated_at"]
     del content["id"]
-
+    del content["is_active"]
     assert response.status_code == status.HTTP_201_CREATED
     assert content == {
         "name": "Iphone 14 Pro Max",
@@ -27,12 +24,10 @@ async def test_controller_get_should_return_success(
     client, products_url, product_inserted
 ):
     response = await client.get(f"{products_url}{product_inserted.id}")
-
     content = response.json()
-
     del content["created_at"]
     del content["updated_at"]
-
+    del content["is_active"]
     assert response.status_code == status.HTTP_200_OK
     assert content == {
         "id": str(product_inserted.id),
@@ -45,7 +40,6 @@ async def test_controller_get_should_return_success(
 
 async def test_controller_get_should_return_not_found(client, products_url):
     response = await client.get(f"{products_url}4fd7cd35-a3a0-4c1f-a78d-d24aa81e7dca")
-
     assert response.status_code == status.HTTP_404_NOT_FOUND
     assert response.json() == {
         "detail": "Product not found with filter: 4fd7cd35-a3a0-4c1f-a78d-d24aa81e7dca"
@@ -55,7 +49,6 @@ async def test_controller_get_should_return_not_found(client, products_url):
 @pytest.mark.usefixtures("products_inserted")
 async def test_controller_query_should_return_success(client, products_url):
     response = await client.get(products_url)
-
     assert response.status_code == status.HTTP_200_OK
     assert isinstance(response.json(), List)
     assert len(response.json()) > 1
@@ -69,10 +62,9 @@ async def test_controller_patch_should_return_success(
     )
 
     content = response.json()
-
     del content["created_at"]
     del content["updated_at"]
-
+    del content["is_active"]
     assert response.status_code == status.HTTP_200_OK
     assert content == {
         "id": str(product_inserted.id),
@@ -87,7 +79,6 @@ async def test_controller_delete_should_return_no_content(
     client, products_url, product_inserted
 ):
     response = await client.delete(f"{products_url}{product_inserted.id}")
-
     assert response.status_code == status.HTTP_204_NO_CONTENT
 
 
@@ -95,8 +86,42 @@ async def test_controller_delete_should_return_not_found(client, products_url):
     response = await client.delete(
         f"{products_url}4fd7cd35-a3a0-4c1f-a78d-d24aa81e7dca"
     )
-
     assert response.status_code == status.HTTP_404_NOT_FOUND
     assert response.json() == {
         "detail": "Product not found with filter: 4fd7cd35-a3a0-4c1f-a78d-d24aa81e7dca"
     }
+
+
+async def test_controller_query_filter_price_should_return_success(
+    client, products_url
+):
+    await client.post(
+        products_url,
+        json={
+            "name": "Iphone barato",
+            "quantity": 1,
+            "price": "4000.00",
+            "status": True,
+        },
+    )
+    await client.post(
+        products_url,
+        json={
+            "name": "Iphone médio",
+            "quantity": 1,
+            "price": "6000.00",
+            "status": True,
+        },
+    )
+    await client.post(
+        products_url,
+        json={"name": "Iphone caro", "quantity": 1, "price": "9000.00", "status": True},
+    )
+
+    response = await client.get(f"{products_url}?min_price=5000&max_price=8000")
+
+    assert response.status_code == 200
+    data = response.json()
+    assert len(data) == 1
+    assert data[0]["name"] == "Iphone médio"
+    assert float(data[0]["price"]) == 6000.00
